moving mount logic
